+++ NEWS.md
@@ -4,6 +4,10 @@
 
 * `find_parameters()` and `get_parameters()` now include the sigma-parameter for non-mixed models from *rstanarm* and *brms* .
 
+## New function
+
+* `get_sigma()` to return the residual standard deviation.
+
 ## Changes to functions
 
 * `format_value()` gains a `zap_small` argument, to prevent scientific printing of numbers if these have more decimal places than indicated by `digits`.